requirements.txt
Flask>=2.0.0
openai==0.28.0
python-dotenv>=0.21.0
gunicorn>=20.1.0 